+++ NEWS.md
@@ -3,6 +3,7 @@
 * describe more fully constructing the datasets table using concept tags
 * added `text_usr()` to add text in normalized device coordinates
 * begin to convert documentation to use markdown via {roxygen2md}
+* `label.ellipse()` re-written to allow a fully general way to specify label positions [Thx: Claude]
 
 ## Version 1.7.9